ExpressionParser: Move FunctionExpression type definitions into another file.
