+++ Source/Core/InputCommon/ControlReference/ExpressionParser.h
@@ -7,7 +7,7 @@
 #include <map>
 #include <memory>
 #include <string>
-#include <utility>
+
 #include "InputCommon/ControllerInterface/Device.h"
 
 namespace ciface::ExpressionParser
